requirements.txt
asgiref==3.2.7
autopep8==1.5.2
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.6
django-rest-framework==0.1.0
djangorestframework==3.11.0
ecdsa==0.15
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pycodestyle==2.5.0
Pygments==2.6.1
python-jose==3.1.0
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
